Dag util for common functions across dags
